Clean up and refactor functions and components
- Merged multiple functions with similar purposes to simplify the code.
- Created a reusable utility class to improve maintainability and consistency across components.
- Reduced code duplication by consolidating common logic into organized utility functions.
- Added comments to improve clarity in frequently used code sections.
- Created new reusable components such as Toolbar (previously ProductListHeader) used in both Shop and Wishlist pages.
- Created a Pagination component (previously ProductListBottom) for use in both Shop and Wishlist pages.
- Added comments to CSS files to improve readability and maintainability.
- Improved code readability and scalability by adopting a more modular structure.
