+++ src/components/Main/main.module.css
@@ -777,7 +777,6 @@ main {
 
 .cart-contents .item-list .item {
  min-height: 125px;
- flex: auto;
  width: 100%;
  display: grid;
  grid-template-columns: min-content 1fr min-content;
@@ -1073,6 +1072,7 @@ main {
  border: 1px solid var(--light-500-color);
  border-radius: 15px;
  overflow: hidden;
+ transition: var(--transition-medium);
 }
 
 .notification-list .notification-item::before {
@@ -1094,6 +1094,10 @@ main {
 
 .notification-list .notification-item.unread::before {
  transform: scaleX(1);
+}
+
+.notification-list .notification-item.open {
+ box-shadow: 0px 2px 4px var(--dark-500-color);
 }
 
 .notification-list .notification-item .time {
@@ -1123,12 +1127,39 @@ main {
 }
 
 .notification-list .notification-item .detail-button {
+ position: relative;
  grid-column: 1 / -1;
  cursor: pointer;
  display: flex;
  align-items: center;
  justify-content: start;
  gap: 4px;
+}
+
+.notification-list .notification-item .detail-button::before {
+ content: "";
+ position: absolute;
+ bottom: 0;
+ left: 0;
+ width: 100%;
+ height: 2px;
+ background: var(--dark-color);
+ transition: 0.3s transform ease, 0.3s opacity ease;
+ transform-origin: right;
+ opacity: 0;
+ transform: scaleX(0);
+}
+
+.notification-list .notification-item .detail-button:hover::before {
+ opacity: 1;
+ transform: scaleX(1);
+ transform-origin: left;
+}
+
+.notification-list .notification-item .detail-button.open::before {
+ opacity: 1;
+ transform: scaleX(1);
+ transform-origin: left;
 }
 
 .notification-list .notification-item .detail-button svg {
@@ -1891,6 +1922,7 @@ main {
 }
 
 @media (max-width: 1300px) {
+ /* Home */
  .categories .category h3 {
   font-size: 1.8rem;
  }
@@ -2005,6 +2037,7 @@ main {
 }
 
 @media (max-width: 950px) and (max-height: 550px) {
+ /* Home */
  .item-display .displayed-image {
   height: 100%;
  }
@@ -2102,6 +2135,7 @@ main {
   min-height: calc(100vh - 90px);
  }
 
+ /* Home */
  .item-display h2 {
   font-size: 0.9rem;
  }
@@ -2138,6 +2172,7 @@ main {
   min-height: calc(100vh - 136px);
  }
 
+ /* Home */
  .hero {
   height: calc(100vh - 150px - 2rem);
  }
@@ -2309,12 +2344,14 @@ main {
 }
 
 @media (max-width: 600px) and (max-height: 750px) {
+ /* Home */
  .item-display h2 {
   font-size: 1.25rem;
  }
 }
 
 @media (max-width: 400px) {
+ /* Home */
  .item-display h2 {
   font-size: 1rem;
  }
@@ -2329,12 +2366,14 @@ main {
 }
 
 @media (max-width: 400px) and (max-height: 700px) {
+ /* Home */
  .item-display .displayed-image {
   height: 70%;
  }
 }
 
 @media (max-height: 400px) {
+ /* Home */
  .categories .category h3 {
   font-size: 1rem;
  }
@@ -2357,6 +2396,7 @@ main {
 }
 
 @media (max-width: 700px) and (max-height: 400px) {
+ /* Home */
  .item-display h2 {
   font-size: 0.8rem;
  }
