+++ src/pages/Cart/Cart.jsx
@@ -12,6 +12,11 @@ function Cart() {
   isExiting,
   setIsExiting,
  } = useOutletContext();
+
+ if (products.length === 0) {
+  return null;
+ }
+
  return (
   <>
    <CartMain
